fix(types): Resolve all pyright errors and increase test coverage to 90%
Type fixes:
- runs.py: Convert Decimal to float for statistics.mean() result
- utils.py: Type-safe access to model_dump via getattr
- utils.py: Suppress intentional TypeVar warning in overload
- conftest.py: Convert float to Decimal for total_cost field
- test_filters.py: Add type ignore for intentional invalid test data

New tests (39 total):
- test_datasets.py: 16 tests for dataset IDs, exclude, count, output,
  get table/JSON, create, push commands
- test_examples.py: 12 tests for exclude, count, output, get with
  fields/as-of, create with metadata/split
- test_prompts.py: 11 tests for exclude, count, output, get with
  commit/fallback, push with tags/public

Coverage improvements:
- datasets.py: 59% → 99%
- examples.py: 67% → 97%
- prompts.py: 69% → 99%
- Overall: 84% → 90%

+++ src/langsmith_cli/commands/runs.py
@@ -291,7 +291,7 @@ def compute_metrics(
             for r in runs
             if hasattr(r, "total_cost") and r.total_cost is not None
         ]
-        result["avg_cost"] = statistics.mean(costs) if costs else 0.0
+        result["avg_cost"] = float(statistics.mean(costs)) if costs else 0.0
 
     return result
